Rename module to ./recursion-detector
